Add support for tapping on Add New... on mobile
On at least Chrome on Android and Safari on iOS, typing a few
characters in a selectize field, then tapping Add new... would result in a
blank field and no behind-the-scenes selection.

This was happening because the event handler responsible for running
onOptionSelect to generate and select an item was only listening to
mousedown and click events. On mobile browsers, the underlying input
field would be blurred (and therefore cleared) whenever a touchstart
event fired outside its bounding box. This was happening before
touchstart in turn called mousedown, so the input field would be cleared
prior to the execution of onOptionSelect. onOptionSelect would run, find
no value on the input field, and set the selection accordingly.

Listening for touchstart events seems to completely fix the problem in
my limited testing.